# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "Documentação da API - API para gerenciamento de produtos",
        "contact": {
            "email": "danielthooth@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista todos os produtos",
                "description": "Api para gerenciamento de produtos",
                "operationId": "e0202514059b097ae6790df19355f578",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Nome do produto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Data de início",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "Data de término",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Quantidade de produtos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Camisa Polo"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Camisa polo branca com estampa azul"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 79.9
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "cost": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 50
                                            },
                                            "status": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "camisa-polo-branca-azul"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado - token ausente ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "message": {
                                                "type": "string",
                                                "example": "Unauthorized"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cria um novo produto",
                "description": "Api para gerenciamento de produtos",
                "operationId": "b1b3ae27ee79dd7a168716f7b73d565f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camisa Polo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Camisa polo branca com estampa azul"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 79.9
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produto criado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Mostra um produto",
                "description": "Api para gerenciamento de produtos",
                "operationId": "ccf839fb1f1b0c3ec939f50c76cd366a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Camisa Polo"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Camisa polo branca com estampa azul"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 79.9
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "cost": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 50
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "camisa-polo-branca-azul"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualiza um produto",
                "description": "Api para gerenciamento de produtos",
                "operationId": "37687c82c4b55acb0b4b235208fc2732",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camisa Polo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Camisa polo branca com estampa azul"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 79.9
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produto atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Deleta um produto",
                "description": "Api para gerenciamento de produtos",
                "operationId": "b5277a733825dcb0bc1ae8464ba78fc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produto deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}/change-status": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Altera o status de um produto para ativo ou inativo",
                "description": "Api para gerenciamento de produtos",
                "operationId": "2004c25a126d393d8842e24b72d91024",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status do produto alterado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Status do produto alterado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}